* The drop target now moves more fluid than before (I hope :) * Fixed the Konqueror plugin to schedule new downloads even if the Knoqueror   window was closed * Modified the startup creation of the Scheduler: we *need* to create it in the   constructor to avoid crashes at KGet startup * Other misc style changes
svn path=/branches/make_kget_cool/kdenetwork/kget/; revision=380768

+++ ui/groupspanel.cpp
@@ -56,7 +56,7 @@ void GroupItem::paintCell(QPainter * p, const QColorGroup & cg, int column, int 
     if(column == 1)
     {
         Group::Info info = group->info();
-        float rectWidth = (width-4) * info.percent / 100;
+        int rectWidth = (int)((width-4) * info.percent / 100);
         
         p->fillRect(2,2,rectWidth, height()-4, cg.brush(QColorGroup::Highlight));
         p->setPen(cg.foreground());
